chore: create separate CashApp framework module
