+++ Adyen/Core/Core Protocols/AnyCashAppPayConfiguration.swift
@@ -4,7 +4,6 @@
 // This file is open source and available under the MIT license. See the LICENSE file for more info.
 //
 
-@_spi(AdyenInternal) import Adyen
 import Foundation
 
 /// Defines the configuration for Cash App Component
