+++ AdyenCashAppPay/AdyenCashAppPay.h
@@ -1,0 +1,19 @@
+//
+//  AdyenCashAppPay.h
+//  AdyenCashAppPay
+//
+//  Created by Eren Besel on 4/11/23.
+//  Copyright © 2023 Adyen. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for AdyenCashAppPay.
+FOUNDATION_EXPORT double AdyenCashAppPayVersionNumber;
+
+//! Project version string for AdyenCashAppPay.
+FOUNDATION_EXPORT const unsigned char AdyenCashAppPayVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <AdyenCashAppPay/PublicHeader.h>
+
+
